Added "api" module for future API implementations.
Moved InputManager to this module as more appropriate.
